Refactor VM to new file
for future use
